==> gadid_mb_aquaria/8_mb_aquarium_plots.Rmd <==
---
title: "metabarcoding plots of tank samples"
author: "Kimberly Ledger"
date: "2024-01-18"
output: html_document
---

clear environment
```{r}
rm(list=ls())
```

```{r}
library(tidyverse)
library(ggplot2)
select <- dplyr::select
rename <- dplyr::rename
```

read in tank metadata 
```{r}
metadata <-  read.csv("/home/kimberly.ledger/gadid_aquaria/gadid_mb_aquaria/TankData_MBmodified.csv") %>% 
  filter(groupID != "T") %>%  #remove tap water controls
  select(!sampleID) %>% #remove sample id because estimates are at the tank-level
  unique() %>%
  rename(tank_ID = groupID)
```

some tank metadata summaries 
```{r}
metadata %>%
  filter(Fish_pres == 1) %>%
  filter(N_spp > 1) %>%
  filter(WP_pres == 1) %>%
  arrange(desc(WP_Fbiomass))
```

read in tank estimates
```{r}
raw_tanks_long <- read.csv("/home/kimberly.ledger/gadid_aquaria/gadid_mb_aquaria/tank_raw_estimates.csv") %>%
 mutate(Species = ifelse(Species == "Boreogadus saida", "AC", Species),
         Species = ifelse(Species == "Gadus chalcogrammus", "WP", Species),
         Species = ifelse(Species == "Gadus macrocephalus", "PC", Species))
mock_tanks_long <- read.csv("/home/kimberly.ledger/gadid_aquaria/gadid_mb_aquaria/tank_mock_estimates.csv") %>%
  mutate(Species = ifelse(Species == "Boreogadus saida", "AC", Species),
         Species = ifelse(Species == "Gadus chalcogrammus", "WP", Species),
         Species = ifelse(Species == "Gadus macrocephalus", "PC", Species))
```

merge raw and mock estimates, and join to tank metadata 
```{r}
raw_tanks_long <- raw_tanks_long %>%
  rename(raw.Mean = Mean,
         raw.SD = SD,
         raw.q.025 = q.025,
         raw.Median = Median,
         raw.q.975 = q.975)
mock_tanks_long <- mock_tanks_long %>%
  rename(mock.Mean = Mean,
         mock.SD = SD,
         mock.q.025 = q.025,
         mock.Median = Median,
         mock.q.975 = q.975)
all <- raw_tanks_long %>%
  left_join(mock_tanks_long)
```

## testing allometric scaling - using 0.65 scaling coefficient based on O'Shea's cod mass:SA ratio

calculate fractional abundance and fractional ASM
```{r}
metadata <- metadata %>%
  #mutate(AC_ind_mass = AC_biomass/AC_N) %>%
  #mutate(PC_ind_mass = PC_biomass/PC_N) %>%
  #mutate(WP_ind_mass = WP_biomass/WP_N) %>%
  mutate(AC_ASM = (AC_biomass^0.65)) %>%   ### this is a simplification of ((total AC biomass^0.65)/N of AC) * N of AC
  mutate(PC_ASM = (PC_biomass^0.65)) %>%
  mutate(WP_ASM = (WP_biomass^0.65)) %>%
  mutate(AC_N = ifelse(is.na(AC_N), 0, AC_N), 
         PC_N = ifelse(is.na(PC_N), 0, PC_N),
         WP_N = ifelse(is.na(WP_N), 0, WP_N),
         Tot_N = AC_N + PC_N + WP_N,
         AC_FN = AC_N/Tot_N,
         PC_FN = PC_N/Tot_N,
         WP_FN = WP_N/Tot_N) %>%
  mutate(AC_ASM = ifelse(is.na(AC_ASM), 0, AC_ASM), 
         PC_ASM = ifelse(is.na(PC_ASM), 0, PC_ASM),
         WP_ASM = ifelse(is.na(WP_ASM), 0, WP_ASM),
         Tot_ASM = AC_ASM + PC_ASM + WP_ASM,
         AC_fASM = AC_ASM/Tot_ASM,
         PC_fASM = PC_ASM/Tot_ASM,
         WP_fASM = WP_ASM/Tot_ASM)
```

## pivot fbiomass and join to metabarcoding reads
join - this removes the few T tanks with eDNA reads
```{r}
all_long_biomass <- metadata %>%
  pivot_longer(cols = c(19:21), names_to = "Species", values_to = "Fbiomass") %>%
  mutate(Species = str_remove(Species, "_Fbiomass")) %>%
  left_join(all, by = c("tank_ID", "Species"))
```

## pivot fractional abundance and joing to metabarcoing reads
```{r}
all_long_abundance <- metadata %>%
  pivot_longer(cols = c(36:38), names_to = "Species", values_to = "F_Abundance") %>%
  mutate(Species = str_remove(Species, "_FN")) %>%
  mutate(F_Abundance = ifelse(is.na(F_Abundance), 0, F_Abundance)) %>%
  left_join(all, by = c("tank_ID", "Species"))
```

## pivot fractional ASM and joing to metabarcoing reads
```{r}
all_long_ASM <- metadata %>%
  pivot_longer(cols = c(40:42), names_to = "Species", values_to = "F_ASM") %>%
  mutate(Species = str_remove(Species, "_fASM")) %>%
  mutate(F_ASM = ifelse(is.na(F_ASM), 0, F_ASM)) %>%
  left_join(all, by = c("tank_ID", "Species"))
```



##extracting data for allometric scaling test
```{r}
temp1 <- metadata %>%
  pivot_longer(cols = c(25:27), names_to = "Species", values_to = "N") %>%
  mutate(Species = str_remove(Species, "_N")) %>%
  left_join(all, by = c("tank_ID", "Species")) %>%
  select(tank_ID, N_spp, PC_pres, Species, N, mock.Mean) %>%
  filter(N_spp > 0) %>%
  #filter(PC_pres == 1) %>%
  unite("Population", tank_ID, Species, remove = F) %>%
  mutate(N = ifelse(is.na(N), 0, N))

temp2 <- metadata %>%
  mutate(AC_mass = AC_biomass/AC_N) %>% 
  mutate(PC_mass = PC_biomass/PC_N) %>%
  mutate(WP_mass = WP_biomass/WP_N) %>%
  filter(N_spp > 0) %>%
  #filter(PC_pres == 1) %>%
  select(tank_ID, AC_mass, PC_mass, WP_mass) %>%
  pivot_longer(cols = c(2:4), names_to = "Species", values_to = "Mass") %>%
  mutate(Species = str_remove(Species, "_mass")) %>%
  unite("Population", tank_ID, Species, remove = F) %>%
  mutate(Mass = ifelse(is.na(Mass), 0, Mass))

#write.csv(temp2, "allometry_mass.csv") 
#write.csv(temp1, "allometry_N.csv")
```


### new figure 3
facet by species and add lines of best fit. 
```{r}
mypal <- c("#6a6599", "#79af97", "#00a1d5")   #purple, green, blue

fig_mock_biomass_by_spp <- all_long_biomass %>%
  filter(Fish_pres == "1") %>%
  filter(n.tank.dissim == "0") %>%
  mutate(Species = ifelse(Species == "AC", "Arctic cod", Species)) %>%
  mutate(Species = ifelse(Species == "PC", "Pacific cod", Species)) %>%
  mutate(Species = ifelse(Species == "WP", "Walleye pollock", Species)) %>%
  ggplot() + 
  geom_errorbar(aes(x=Fbiomass,ymin=mock.q.025,ymax=mock.q.975,color=Species),width=0, size = .5) +
  geom_point(aes(x=Fbiomass,y=mock.Mean,color=Species), size = 2) +
  #geom_smooth(aes(x=Fbiomass,y=mock.Mean,color=Species), method = "lm", formula = y ~ x - 1, se = FALSE) +
  geom_smooth(aes(x=Fbiomass,y=mock.Mean,color=Species), method = "lm", formula = y ~ x, se = FALSE) +
  geom_abline(intercept = 0, slope = 1, linetype = "dashed", color = "black")  + 
  facet_wrap(~Species) +
  scale_color_manual(values = mypal) + 
  theme_bw() + 
    labs(x = "proportional biomass",
       y = "corrected eDNA read proportion") +
  theme(axis.text = element_text(size = 10),
        axis.title = element_text(size = 14)) + 
        #legend.text = element_text(size = 12),
        #legend.title = element_text(size = 14)
  guides(color = FALSE)
fig_mock_biomass_by_spp
```

facet by species and color a few of the tanks  
```{r}
fig_mock_biomass_by_spp_coloredtanks <- all_long_biomass %>%
  filter(Fish_pres == "1") %>%
  filter(n.tank.dissim == "0") %>%
  mutate(Species = ifelse(Species == "AC", "Arctic cod", Species)) %>%
  mutate(Species = ifelse(Species == "PC", "Pacific cod", Species)) %>%
  mutate(Species = ifelse(Species == "WP", "Walleye pollock", Species)) %>%
  mutate(to_color = ifelse(tank_ID %in% c("3", "46", "2", "10", "71", "80", "89"), tank_ID, NA)) %>%
  ggplot() + 
  geom_errorbar(aes(x=Fbiomass,ymin=mock.q.025,ymax=mock.q.975,color=to_color),width=0, size = .5) +
  geom_point(aes(x=Fbiomass,y=mock.Mean,color=to_color), size = 2) +
  #geom_smooth(aes(x=Fbiomass,y=mock.Mean,color=Species), method = "lm", formula = y ~ x - 1, se = FALSE) +
  #geom_smooth(aes(x=Fbiomass,y=mock.Mean), method = "lm", formula = y ~ x, se = FALSE, color = "light grey") +
  geom_abline(intercept = 0, slope = 1, linetype = "dashed", color = "black")  + 
  facet_wrap(~Species) +
  #scale_color_manual(values = mypal) + 
  theme_bw() + 
    labs(x = "proportional biomass",
       y = "corrected eDNA read proportion",
       color = "tank ID") +
  theme(axis.text = element_text(size = 10),
        axis.title = element_text(size = 14)) #+ 
        #legend.text = element_text(size = 12),
        #legend.title = element_text(size = 14)
  #guides(color = FALSE)
fig_mock_biomass_by_spp_coloredtanks
```



## try coloring by tank 
color by tank ID
```{r}
fig_mock_biomass_by_spp_coloredtanks <- all_long_biomass %>%
  filter(Fish_pres == "1") %>%
  filter(n.tank.dissim == "0") %>%
  mutate(Species = ifelse(Species == "AC", "Arctic cod", Species)) %>%
  mutate(Species = ifelse(Species == "PC", "Pacific cod", Species)) %>%
  mutate(Species = ifelse(Species == "WP", "Walleye pollock", Species)) %>%
  ggplot() + 
  geom_errorbar(aes(x=Fbiomass,ymin=mock.q.025,ymax=mock.q.975,color=tank_ID),width=0, size = .5) +
  geom_point(aes(x=Fbiomass,y=mock.Mean,color=tank_ID), size = 2) +
  #geom_smooth(aes(x=Fbiomass,y=mock.Mean,color=Species), method = "lm", formula = y ~ x - 1, se = FALSE) +
  geom_smooth(aes(x=Fbiomass,y=mock.Mean), method = "lm", formula = y ~ x, se = FALSE) +
  geom_abline(intercept = 0, slope = 1, linetype = "dashed", color = "black")  + 
  facet_wrap(~Species) +
  #scale_color_manual(values = mypal) + 
  theme_bw() + 
    labs(x = "proportional biomass",
       y = "corrected eDNA read proportion") +
  theme(axis.text = element_text(size = 10),
        axis.title = element_text(size = 14)) #+ 
        #legend.text = element_text(size = 12),
        #legend.title = element_text(size = 14)
  #guides(color = FALSE)
fig_mock_biomass_by_spp_coloredtanks
```


```{r}
#ggsave("figures/biomass_by_tank.png", plot = fig_mock_biomass_by_spp_coloredtanks, width = 8.5, height = 3.5, dpi = 300)
```


### new figure S3A 
and raw proportions 
```{r}
fig_raw_biomass_by_spp <- all_long_biomass %>%
  filter(Fish_pres == "1") %>%
  filter(n.tank.dissim == "0") %>%
  mutate(Species = ifelse(Species == "AC", "Arctic cod", Species)) %>%
  mutate(Species = ifelse(Species == "PC", "Pacific cod", Species)) %>%
  mutate(Species = ifelse(Species == "WP", "Walleye pollock", Species)) %>%
  ggplot() + 
  geom_errorbar(aes(x=Fbiomass,ymin=raw.q.025,ymax=raw.q.975,color=Species),width=0, size = .5) +
  geom_point(aes(x=Fbiomass,y=raw.Mean,color=Species), size = 2) +
  #geom_smooth(aes(x=Fbiomass,y=mock.Mean,color=Species), method = "lm", formula = y ~ x - 1, se = FALSE) +
  geom_smooth(aes(x=Fbiomass,y=raw.Mean,color=Species), method = "lm", formula = y ~ x, se = FALSE) +
  geom_abline(intercept = 0, slope = 1, linetype = "dashed", color = "black")  + 
  facet_wrap(~Species) +
  scale_color_manual(values = mypal) + 
  theme_bw() + 
    labs(x = "proportional biomass",
       y = "uncorrected eDNA read proportion") +
  theme(axis.text = element_text(size = 10),
        axis.title = element_text(size = 14)) + 
        #legend.text = element_text(size = 12),
        #legend.title = element_text(size = 14)
  guides(color = FALSE)
fig_raw_biomass_by_spp
```


```{r}
#ggsave("figures/raw_read_biomass_by_spp.png", plot = fig_raw_biomass_by_spp, width = 8.5, height = 3.5, dpi = 300)
#ggsave("figures/mock_read_biomass_by_spp.png", plot = fig_mock_biomass_by_spp, width = 8.5, height = 3.5, dpi = 300)
```


linear regressions (corrected & uncorrected read proportions vs biomass) for each species
- arctic cod 
```{r}
ac_biomass <- all_long_biomass %>%
  filter(n.tank.dissim != "2") %>%
  filter(Fish_pres == "1") %>%
  filter(Species == "AC") 

raw_mod_ac_biomass <- lm(raw.Mean ~ Fbiomass, data = ac_biomass)
mock_mod_ac_biomass <- lm(mock.Mean ~ Fbiomass, data = ac_biomass)
summary(raw_mod_ac_biomass)
summary(mock_mod_ac_biomass)
```

- pacific cod
```{r}
pc_biomass <- all_long_biomass %>%
  filter(n.tank.dissim != "2") %>%
  filter(Fish_pres == "1") %>%
  filter(Species == "PC") 

raw_mod_pc_biomass <- lm(raw.Mean ~ Fbiomass, data = pc_biomass)
mock_mod_pc_biomass <- lm(mock.Mean ~ Fbiomass, data = pc_biomass)
summary(raw_mod_pc_biomass)
summary(mock_mod_pc_biomass)
```

- walleye pollock
```{r}
wp_biomass <- all_long_biomass %>%
  filter(n.tank.dissim != "2") %>%
  filter(Fish_pres == "1") %>%
  filter(Species == "WP") 

raw_mod_wp_biomass <- lm(raw.Mean ~ Fbiomass, data = wp_biomass)
mock_mod_wp_biomass <- lm(mock.Mean ~ Fbiomass, data = wp_biomass)
summary(raw_mod_wp_biomass)
summary(mock_mod_wp_biomass)
```


## now for abundance 

- new figure S4A 
```{r}
mypal <- c("#6a6599", "#79af97", "#00a1d5")   #purple, green, blue

fig_mock_abund_by_spp <- all_long_abundance %>%
  filter(Fish_pres == "1") %>%
  filter(n.tank.dissim == "0") %>%
  mutate(Species = ifelse(Species == "AC", "Arctic cod", Species)) %>%
  mutate(Species = ifelse(Species == "PC", "Pacific cod", Species)) %>%
  mutate(Species = ifelse(Species == "WP", "Walleye pollock", Species)) %>%
  ggplot() + 
  geom_errorbar(aes(x=F_Abundance,ymin=mock.q.025,ymax=mock.q.975,color=Species),width=0, size = .5) +
  geom_point(aes(x=F_Abundance,y=mock.Mean,color=Species), size = 2) +
  #geom_smooth(aes(x=F_Abundance,y=mock.Mean,color=Species), method = "lm", formula = y ~ x - 1, se = FALSE) +
  geom_smooth(aes(x=F_Abundance,y=mock.Mean,color=Species), method = "lm", formula = y ~ x, se = FALSE) +
  geom_abline(intercept = 0, slope = 1, linetype = "dashed", color = "black")  + 
  facet_wrap(~Species) +
  scale_color_manual(values = mypal) + 
  theme_bw() + 
    labs(x = "proportional abundance",
       y = "corrected eDNA read proportion") +
  theme(axis.text = element_text(size = 10),
        axis.title = element_text(size = 14)) + 
        #legend.text = element_text(size = 12),
        #legend.title = element_text(size = 14)
  guides(color = FALSE)
fig_mock_abund_by_spp
```

- new figure S4B
```{r}
mypal <- c("#6a6599", "#79af97", "#00a1d5")   #purple, green, blue

fig_raw_abund_by_spp <- all_long_abundance %>%
  filter(Fish_pres == "1") %>%
  filter(n.tank.dissim == "0") %>%
  mutate(Species = ifelse(Species == "AC", "Arctic cod", Species)) %>%
  mutate(Species = ifelse(Species == "PC", "Pacific cod", Species)) %>%
  mutate(Species = ifelse(Species == "WP", "Walleye pollock", Species)) %>%
  ggplot() + 
  geom_errorbar(aes(x=F_Abundance,ymin=raw.q.025,ymax=raw.q.975,color=Species),width=0, size = .5) +
  geom_point(aes(x=F_Abundance,y=raw.Mean,color=Species), size = 2) +
  #geom_smooth(aes(x=F_Abundance,y=raw.Mean,color=Species), method = "lm", formula = y ~ x - 1, se = FALSE) +
  geom_smooth(aes(x=F_Abundance,y=raw.Mean,color=Species), method = "lm", formula = y ~ x, se = FALSE) +
  geom_abline(intercept = 0, slope = 1, linetype = "dashed", color = "black")  + 
  facet_wrap(~Species) +
  scale_color_manual(values = mypal) + 
  theme_bw() + 
    labs(x = "proportional abundance",
       y = "uncorrected eDNA read proportion") +
  theme(axis.text = element_text(size = 10),
        axis.title = element_text(size = 14)) + 
        #legend.text = element_text(size = 12),
        #legend.title = element_text(size = 14)
  guides(color = FALSE)
fig_raw_abund_by_spp
```

put together into one plot 
```{r, fig.width=10, fig.height=4}
library(patchwork)

abundance_fig <- fig_mock_abund_by_spp / fig_raw_abund_by_spp + plot_layout(axes = 'collect') + plot_annotation(tag_levels = 'A') &
    theme(plot.tag = element_text(face = 'bold'))
abundance_fig
```


```{r}
#ggsave("figures/mock_raw_read_abundance_by_spp.png", plot = abundance_fig, width = 8.5, height = 7, dpi = 300)
```


now for the models w/ abundance 

linear regressions (corrected & uncorrected read proportions vs biomass) for each species
- arctic cod 
```{r}
ac_abund <- all_long_abundance %>%
  filter(n.tank.dissim != "2") %>%
  filter(Fish_pres == "1") %>%
  filter(Species == "AC") 

raw_mod_ac_abund <- lm(raw.Mean ~ F_Abundance, data = ac_abund)
mock_mod_ac_abund <- lm(mock.Mean ~ F_Abundance, data = ac_abund)
summary(raw_mod_ac_abund)
summary(mock_mod_ac_abund)
```

- pacific cod
```{r}
pc_abund <- all_long_abundance %>%
  filter(n.tank.dissim != "2") %>%
  filter(Fish_pres == "1") %>%
  filter(Species == "PC") 

raw_mod_pc_abund <- lm(raw.Mean ~ F_Abundance, data = pc_abund)
mock_mod_pc_abund <- lm(mock.Mean ~ F_Abundance, data = pc_abund)
summary(raw_mod_pc_abund)
summary(mock_mod_pc_abund)
```

- walleye pollock
```{r}
wp_abund <- all_long_abundance %>%
  filter(n.tank.dissim != "2") %>%
  filter(Fish_pres == "1") %>%
  filter(Species == "WP") 

raw_mod_wp_abund <- lm(raw.Mean ~ F_Abundance, data = wp_abund)
mock_mod_wp_abund <- lm(mock.Mean ~ F_Abundance, data = wp_abund)
summary(raw_mod_wp_abund)
summary(mock_mod_wp_abund)
```


## now trying ASM
- new figure S5A 
```{r}
mypal <- c("#6a6599", "#79af97", "#00a1d5")   #purple, green, blue

fig_mock_ASM_by_spp <- all_long_ASM %>%
  filter(Fish_pres == "1") %>%
  filter(n.tank.dissim == "0") %>%
  mutate(Species = ifelse(Species == "AC", "Arctic cod", Species)) %>%
  mutate(Species = ifelse(Species == "PC", "Pacific cod", Species)) %>%
  mutate(Species = ifelse(Species == "WP", "Walleye pollock", Species)) %>%
  ggplot() + 
  geom_errorbar(aes(x=F_ASM,ymin=mock.q.025,ymax=mock.q.975,color=Species),width=0, size = .5) +
  geom_point(aes(x=F_ASM,y=mock.Mean,color=Species), size = 2) +
  #geom_smooth(aes(x=F_ASM,y=mock.Mean,color=Species), method = "lm", formula = y ~ x - 1, se = FALSE) +
  geom_smooth(aes(x=F_ASM,y=mock.Mean,color=Species), method = "lm", formula = y ~ x, se = FALSE) +
  geom_abline(intercept = 0, slope = 1, linetype = "dashed", color = "black")  + 
  facet_wrap(~Species) +
  scale_color_manual(values = mypal) + 
  theme_bw() + 
    labs(x = "proportional ASM",
       y = "corrected eDNA read proportion") +
  theme(axis.text = element_text(size = 10),
        axis.title = element_text(size = 14)) + 
        #legend.text = element_text(size = 12),
        #legend.title = element_text(size = 14)
  guides(color = FALSE)
fig_mock_ASM_by_spp
```

- new figure S5B
```{r}
mypal <- c("#6a6599", "#79af97", "#00a1d5")   #purple, green, blue

fig_raw_ASM_by_spp <- all_long_ASM %>%
  filter(Fish_pres == "1") %>%
  filter(n.tank.dissim == "0") %>%
  mutate(Species = ifelse(Species == "AC", "Arctic cod", Species)) %>%
  mutate(Species = ifelse(Species == "PC", "Pacific cod", Species)) %>%
  mutate(Species = ifelse(Species == "WP", "Walleye pollock", Species)) %>%
  ggplot() + 
  geom_errorbar(aes(x=F_ASM,ymin=raw.q.025,ymax=raw.q.975,color=Species),width=0, size = .5) +
  geom_point(aes(x=F_ASM,y=raw.Mean,color=Species), size = 2) +
  #geom_smooth(aes(x=F_ASM,y=raw.Mean,color=Species), method = "lm", formula = y ~ x - 1, se = FALSE) +
  geom_smooth(aes(x=F_ASM,y=raw.Mean,color=Species), method = "lm", formula = y ~ x, se = FALSE) +
  geom_abline(intercept = 0, slope = 1, linetype = "dashed", color = "black")  + 
  facet_wrap(~Species) +
  scale_color_manual(values = mypal) + 
  theme_bw() + 
    labs(x = "proportional ASM",
       y = "uncorrected eDNA read proportion") +
  theme(axis.text = element_text(size = 10),
        axis.title = element_text(size = 14)) + 
        #legend.text = element_text(size = 12),
        #legend.title = element_text(size = 14)
  guides(color = FALSE)
fig_raw_ASM_by_spp
```

put together into one plot 
```{r, fig.width=10, fig.height=4}
library(patchwork)

ASM_fig <- fig_mock_ASM_by_spp / fig_raw_ASM_by_spp + plot_layout(axes = 'collect') + plot_annotation(tag_levels = 'A') &
    theme(plot.tag = element_text(face = 'bold'))
ASM_fig
```


```{r}
ggsave("figures/mock_raw_read_ASM_by_spp.png", plot = ASM_fig, width = 8.5, height = 7, dpi = 300)
```


now for the models w/ ASM 

linear regressions (corrected & uncorrected read proportions vs ASM) for each species
- arctic cod 
```{r}
ac_ASM <- all_long_ASM %>%
  filter(n.tank.dissim != "2") %>%
  filter(Fish_pres == "1") %>%
  filter(Species == "AC") 

raw_mod_ac_ASM <- lm(raw.Mean ~ F_ASM, data = ac_ASM)
mock_mod_ac_ASM <- lm(mock.Mean ~ F_ASM, data = ac_ASM)
summary(raw_mod_ac_ASM)
summary(mock_mod_ac_ASM)
```

- pacific cod
```{r}
pc_ASM <- all_long_ASM %>%
  filter(n.tank.dissim != "2") %>%
  filter(Fish_pres == "1") %>%
  filter(Species == "PC") 

raw_mod_pc_ASM <- lm(raw.Mean ~ F_ASM, data = pc_ASM)
mock_mod_pc_ASM <- lm(mock.Mean ~ F_ASM, data = pc_ASM)
summary(raw_mod_pc_ASM)
summary(mock_mod_pc_ASM)
```

- walleye pollock
```{r}
wp_ASM <- all_long_ASM %>%
  filter(n.tank.dissim != "2") %>%
  filter(Fish_pres == "1") %>%
  filter(Species == "WP") 

raw_mod_wp_ASM <- lm(raw.Mean ~ F_ASM, data = wp_ASM)
mock_mod_wp_ASM <- lm(mock.Mean ~ F_ASM, data = wp_ASM)
summary(raw_mod_wp_ASM)
summary(mock_mod_wp_ASM)
```


## compare AIC of models
```{r}
AIC(mock_mod_ac_biomass, mock_mod_ac_ASM, mock_mod_ac_abund)
AIC(mock_mod_pc_biomass, mock_mod_pc_ASM, mock_mod_pc_abund)
AIC(mock_mod_wp_biomass, mock_mod_wp_ASM, mock_mod_wp_abund)
```

```{r}
AIC(raw_mod_ac_biomass, raw_mod_ac_ASM, raw_mod_ac_abund)
AIC(raw_mod_pc_biomass, raw_mod_pc_ASM, raw_mod_pc_abund)
AIC(raw_mod_wp_biomass, raw_mod_wp_ASM, raw_mod_wp_abund)
```


### pcod - individual analysis... 
make individual plots for each species on their own - using 1:1 line 
```{r}
fig_abund_PC_age <- all_long_abundance %>%
  filter(n.tank.dissim == "0") %>%
  filter(Fish_pres == "1") %>%
  mutate(Species = ifelse(Species == "AC", "Arctic cod", Species)) %>%
  mutate(Species = ifelse(Species == "PC", "Pacific cod", Species)) %>%
  mutate(Species = ifelse(Species == "WP", "Walleye pollock", Species)) %>%
  filter(Species == "Pacific cod") %>%
  ggplot() + 
  geom_errorbar(aes(x=F_Abundance,ymin=mock.q.025,ymax=mock.q.975,color=PC_age),width=0) +
  geom_point(aes(x=F_Abundance,y=mock.Mean,color=PC_age), size = 3) + 
  #geom_errorbar(aes(x=F_Abundance,ymin=q.025,ymax=q.975,color=Species),width=0, size = 0.5) +
  #geom_point(aes(x=F_Abundance,y=Mean,color=Species), size = 3) +
  #geom_smooth(aes(x=F_Abundance,y=Mean,color=Species,fill=Species), method = "lm", se = FALSE, alpha = 0.5) +
  #geom_smooth(aes(x=F_Abundance,y=Mean), color = "black", linetype = "dashed",method = "lm", se = FALSE) +
  geom_abline(intercept = 0, slope = 1, linetype = "dashed", color = "black")  + 
  #scale_color_material("green") + 
  #scale_color_manual(values = "#79af97") +   #"#6a6599", "#79af97", "#00a1d5")   #purple, green, blue
  theme_classic() + 
  labs(x = "proportional abundance",
       y = "corrected eDNA read proportion",
       color = "Pacific cod age") +
  theme(axis.text = element_text(size = 12),
        axis.title = element_text(size = 14),
        legend.text = element_text(size = 12),
        legend.title = element_text(size = 14))
fig_abund_PC_age
```

```{r}
#ggsave("figures/read_abund_PC_age.png", plot = fig_abund_PC_age, width = 8.5, height = 6, dpi = 300)
```


plot abundance - with biomass scale 
```{r}
pc_abund <- all_long_abundance %>%
  filter(n.tank.dissim == "0") %>%
  filter(Fish_pres == "1") %>%
  mutate(AC_ind_mass = AC_biomass/AC_N) %>%
  mutate(PC_ind_mass = PC_biomass/PC_N) %>%
  mutate(WP_ind_mass = WP_biomass/WP_N) %>%
  select(F_Abundance, mock.Mean, mock.q.025, mock.q.975, Species, AC_ind_mass, PC_ind_mass, WP_ind_mass) %>%
  filter(F_Abundance > 0) %>%
  pivot_longer(cols = c(6:8), values_to = "biomass_per_N") %>%
  mutate(name = ifelse(name == "AC_ind_mass", "AC", name)) %>%
  mutate(name = ifelse(name == "PC_ind_mass", "PC", name)) %>%
  mutate(name = ifelse(name == "WP_ind_mass", "WP", name)) %>%
  filter(Species == name) %>%
  select(!name)
  
fig_abund_w_ind_biomass <- pc_abund %>%  
  mutate(Species = ifelse(Species == "AC", "Arctic cod", Species)) %>%
  mutate(Species = ifelse(Species == "PC", "Pacific cod", Species)) %>%
  mutate(Species = ifelse(Species == "WP", "Walleye pollock", Species)) %>%
  filter(Species == "Pacific cod") %>%
  ggplot() + 
  geom_errorbar(aes(x=F_Abundance,ymin=mock.q.025,ymax=mock.q.975,color=Species),width=0, size = 0.5) +
  geom_point(aes(x=F_Abundance,y=mock.Mean,color=Species, fill=biomass_per_N), shape = 21, size = 2, stroke = 1) +
  #geom_smooth(aes(x=F_Abundance,y=Mean), color = "black", linetype = "dashed",method = "lm", se = FALSE) +
  geom_abline(intercept = 0, slope = 1, linetype = "dashed", color = "black")  + 
  scale_color_manual(values = "#79af97") +   #"#6a6599", "#79af97", "#00a1d5"   #purple, green, blue
  #scale_color_manual(values = mypal) + 
  #scale_fill_continuous(low = "white", high = "black", trans = "log10") +  # Apply log transformation to fill scale
  scale_fill_continuous(low = "white", high = "black", trans = "sqrt") +  # Apply log transformation to fill scale
  theme_classic() + 
  labs(x = "proportional abundance",
       y = "corrected eDNA read proportion",
       fill = "grams per individual") +
  scale_x_continuous(limits = c(0, 1)) +
  scale_y_continuous(limits = c(0, 1)) + 
  #theme(axis.text = element_text(size = 12),
  #      axis.title = element_text(size = 14),
  #      legend.text = element_text(size = 12),
  #      legend.title = element_text(size = 14)) + 
  guides(color = "none")
fig_abund_w_ind_biomass
```

```{r}
#ggsave("figures/read_abund_PCOD_biomass.png", plot = fig_abund_w_ind_biomass, width = 8.5, height = 6, dpi = 300)
```


plot ASM - with biomass scale 
```{r}
pc_ASM <- all_long_ASM %>%
  filter(n.tank.dissim == "0") %>%
  filter(Fish_pres == "1") %>%
  mutate(AC_ind_mass = AC_biomass/AC_N) %>%
  mutate(PC_ind_mass = PC_biomass/PC_N) %>%
  mutate(WP_ind_mass = WP_biomass/WP_N) %>%
  select(F_ASM, mock.Mean, mock.q.025, mock.q.975, Species, AC_ind_mass, PC_ind_mass, WP_ind_mass) %>%
  filter(F_ASM > 0) %>%
  pivot_longer(cols = c(6:8), values_to = "biomass_per_N") %>%
  mutate(name = ifelse(name == "AC_ind_mass", "AC", name)) %>%
  mutate(name = ifelse(name == "PC_ind_mass", "PC", name)) %>%
  mutate(name = ifelse(name == "WP_ind_mass", "WP", name)) %>%
  filter(Species == name) %>%
  select(!name)
  
fig_ASM_w_ind_biomass <- pc_ASM %>%  
  mutate(Species = ifelse(Species == "AC", "Arctic cod", Species)) %>%
  mutate(Species = ifelse(Species == "PC", "Pacific cod", Species)) %>%
  mutate(Species = ifelse(Species == "WP", "Walleye pollock", Species)) %>%
  filter(Species == "Pacific cod") %>%
  ggplot() + 
  geom_errorbar(aes(x=F_ASM,ymin=mock.q.025,ymax=mock.q.975,color=Species),width=0, size = 0.5) +
  geom_point(aes(x=F_ASM,y=mock.Mean,color=Species, fill=biomass_per_N), shape = 21, size = 2, stroke = 1) +
  #geom_smooth(aes(x=F_Abundance,y=Mean), color = "black", linetype = "dashed",method = "lm", se = FALSE) +
  geom_abline(intercept = 0, slope = 1, linetype = "dashed", color = "black")  + 
  scale_color_manual(values = "#79af97") +   #"#6a6599", "#79af97", "#00a1d5"   #purple, green, blue
  #scale_color_manual(values = mypal) + 
  #scale_fill_continuous(low = "white", high = "black", trans = "log10") +  # Apply log transformation to fill scale
  scale_fill_continuous(low = "white", high = "black", trans = "sqrt") +  # Apply log transformation to fill scale
  theme_classic() + 
  labs(x = "proportional ASM",
       y = "corrected eDNA read proportion",
       fill = "grams per individual") +
  scale_x_continuous(limits = c(0, 1)) +
  scale_y_continuous(limits = c(0, 1)) + 
  #theme(axis.text = element_text(size = 12),
  #      axis.title = element_text(size = 14),
  #      legend.text = element_text(size = 12),
  #      legend.title = element_text(size = 14)) + 
  guides(color = "none")
fig_ASM_w_ind_biomass
```

plot biomass - with biomass scale 
```{r}
pc_biomass <- all_long_biomass %>%
  filter(n.tank.dissim == "0") %>%
  filter(Fish_pres == "1") %>%
  mutate(AC_ind_mass = AC_biomass/AC_N) %>%
  mutate(PC_ind_mass = PC_biomass/PC_N) %>%
  mutate(WP_ind_mass = WP_biomass/WP_N) %>%
  select(Fbiomass, mock.Mean, mock.q.025, mock.q.975, Species, AC_ind_mass, PC_ind_mass, WP_ind_mass) %>%
  filter(Fbiomass > 0) %>%
  pivot_longer(cols = c(6:8), values_to = "biomass_per_N") %>%
  mutate(name = ifelse(name == "AC_ind_mass", "AC", name)) %>%
  mutate(name = ifelse(name == "PC_ind_mass", "PC", name)) %>%
  mutate(name = ifelse(name == "WP_ind_mass", "WP", name)) %>%
  filter(Species == name) %>%
  select(!name)
  
fig_biomass_w_ind_biomass <- pc_biomass %>%  
  mutate(Species = ifelse(Species == "AC", "Arctic cod", Species)) %>%
  mutate(Species = ifelse(Species == "PC", "Pacific cod", Species)) %>%
  mutate(Species = ifelse(Species == "WP", "Walleye pollock", Species)) %>%
  filter(Species == "Pacific cod") %>%
  ggplot() + 
  geom_errorbar(aes(x=Fbiomass,ymin=mock.q.025,ymax=mock.q.975,color=Species),width=0, size = 0.5) +
  geom_point(aes(x=Fbiomass,y=mock.Mean,color=Species, fill=biomass_per_N), shape = 21, size = 2, stroke = 1) +
  #geom_smooth(aes(x=F_Abundance,y=Mean), color = "black", linetype = "dashed",method = "lm", se = FALSE) +
  geom_abline(intercept = 0, slope = 1, linetype = "dashed", color = "black")  + 
  scale_color_manual(values = "#79af97") +   #"#6a6599", "#79af97", "#00a1d5"   #purple, green, blue
  #scale_color_manual(values = mypal) + 
  #scale_fill_continuous(low = "white", high = "black", trans = "log10") +  # Apply log transformation to fill scale
  scale_fill_continuous(low = "white", high = "black", trans = "sqrt") +  # Apply log transformation to fill scale
  theme_classic() + 
  labs(x = "proportional biomass",
       y = "corrected eDNA read proportion",
       fill = "grams per individual") +
  scale_x_continuous(limits = c(0, 1)) +
  scale_y_continuous(limits = c(0, 1)) + 
  #theme(axis.text = element_text(size = 12),
  #      axis.title = element_text(size = 14),
  #      legend.text = element_text(size = 12),
  #      legend.title = element_text(size = 14)) + 
  guides(color = "none")
fig_biomass_w_ind_biomass
```

now explore the relationship between mass and the read proportion v abundance proportion residuals 

```{r}
#ggsave("figures/Pcod_proportions.png", plot = pc_fig, width = 12, height = 4, dpi = 300)
```


having to input some file to get this to run

read in tank estimates
```{r}
tanks <- read.csv("/home/kimberly.ledger/gadid_metabarcoding/aquaria_DBO/tank_mock_estimates.csv") %>%
  mutate(Species = ifelse(Species == "Boreogadus saida", "AC", Species),
         Species = ifelse(Species == "Gadus chalcogrammus", "WP", Species),
         Species = ifelse(Species == "Gadus macrocephalus", "PC", Species)) %>%
  pivot_wider(names_from = Species, values_from = c(6:9))
```

read in tank metadata 
```{r}
metadata <-  read.csv("/home/kimberly.ledger/gadid_metabarcoding/aquaria_DBO/TankData_MBmodified.csv") %>% 
  filter(groupID != "T") %>%  #remove tap water controls
  select(!sampleID) %>% #remove sample id because estimates are at the tank-level
  unique() %>%
  rename(tank_ID = groupID)
```

join - this removes the few T tanks with eDNA reads. okay for now but will need to report those somewhere... 
```{r}
all <- metadata %>%
  left_join(tanks, by = "tank_ID")
```

```{r}
all$n.tank.dissim <- as.factor(all$n.tank.dissim)
all$n.sample.dissim <- as.factor(all$n.sample.dissim)
```

```{r}
all <- all %>%
  mutate(AC_N = ifelse(is.na(AC_N), 0, AC_N), 
         PC_N = ifelse(is.na(PC_N), 0, PC_N),
         WP_N = ifelse(is.na(WP_N), 0, WP_N),
         Tot_N = AC_N + PC_N + WP_N,
         AC_FN = AC_N/Tot_N,
         PC_FN = PC_N/Tot_N,
         WP_FN = WP_N/Tot_N)
```

compute the perpendicular distance of each point to the line
```{r}
dist_data <- all %>%
  filter(n.tank.dissim == "0") %>%
  filter(Fish_pres == "1") %>%
  mutate(AC_grams_per_N = AC_biomass/AC_N) %>%
  mutate(PC_grams_per_N = PC_biomass/PC_N) %>%
  mutate(WP_grams_per_N = WP_biomass/WP_N)

# Calculate distances from points to the 1:1 line
dist_data$PC_distances_abundance <- dist_data$Mean_PC - dist_data$PC_FN
dist_data$PC_distances_biomass <- dist_data$Mean_PC - dist_data$PC_Fbiomass
dist_data$AC_distances_abundance <- dist_data$Mean_AC - dist_data$AC_FN
dist_data$AC_distances_biomass <- dist_data$Mean_AC - dist_data$AC_Fbiomass
dist_data$WP_distances_abundance <- dist_data$Mean_WP - dist_data$WP_FN
dist_data$WP_distances_biomass <- dist_data$Mean_WP - dist_data$WP_Fbiomass

#dist_data <- dist_data %>%
#  filter(!is.na(PC_age))
dist_data
```


compute the perpendicular distance of each point to the line
```{r}
# Calculate distances from points to the 1:1 line
dist_data$distances_abundance <- dist_data$Mean_PC - dist_data$PC_FN
dist_data$distances_biomass <- dist_data$Mean_PC - dist_data$PC_Fbiomass

dist_data <- dist_data %>%
  filter(!is.na(PC_age))
```


```{r}
dist_data <- dist_data %>%
  mutate(PC_age_group = ifelse(PC_age == 0, "age 0", NA),
         PC_age_group = ifelse(PC_age == 1, "age 1", PC_age_group),
         PC_age_group = ifelse(PC_age == 1.5, "age 1", PC_age_group),
         PC_age_group = ifelse(PC_age == 2, "age 2", PC_age_group),
         PC_age_group = ifelse(PC_age == 7, "age 7", PC_age_group))
```


```{r}
pc_residual_abund <- ggplot(dist_data, aes(y = PC_age_group, x = distances_abundance, fill = PC_age)) + 
  geom_point(shape = 21, cex = 2) + 
  theme_bw() + 
  scale_fill_material("green") +
  geom_vline(xintercept = 0, linetype = "dashed", color = "black") +
  labs(x = "residual (abundance)",
       y = "Pacific cod age") +
  theme(legend.position = "none")+
  xlim(c(-0.5,0.5))
pc_residual_abund
```

```{r}
pc_residual_biomass<- ggplot(dist_data, aes(y = PC_age_group, x = distances_biomass, fill = PC_age)) + 
  geom_point(shape = 21, cex = 2) + 
  theme_bw() + 
  scale_fill_material("green") +
  geom_vline(xintercept = 0, linetype = "dashed", color = "black") +
  labs(x = "residual (biomass)",
       y = "Pacific cod age") +
  theme(legend.position = "none")+
  xlim(c(-0.5,0.5))
pc_residual_biomass
```



now plot mass instead of age 
```{r}
pc_residual_abund_by_mass <- ggplot(dist_data, aes(y = PC_grams_per_N, x = distances_abundance, fill = PC_age)) + 
  geom_point(shape = 21, cex = 2) + 
  theme_bw() + 
  scale_fill_material("green") +
  geom_vline(xintercept = 0, linetype = "dashed", color = "black") +
  scale_y_sqrt() +
  labs(x = "residual (abundance)",
       y = "Pacific cod mass (g)",
       fill = "Pacific cod age") +
  theme(legend.position = "right")+
  xlim(c(-0.5,0.5))
pc_residual_abund_by_mass
```

```{r}
pc_residual_biomass_by_mass <- ggplot(dist_data, aes(y = PC_grams_per_N, x = distances_biomass, fill = PC_age)) + 
  geom_point(shape = 21, cex = 2) + 
  theme_bw() + 
  scale_fill_material("green") +
  geom_vline(xintercept = 0, linetype = "dashed", color = "black") +
  scale_y_sqrt() +
  labs(x = "residual (biomass)",
       y = "Pacific cod mass (g)",
       fill = "Pacific cod age") +
  theme(legend.position = "right")+
  xlim(c(-0.5,0.5))
pc_residual_biomass_by_mass
```

combined figure 
```{r}
prop_fig <- fig_biomass_w_ind_biomass + fig_abund_w_ind_biomass + plot_layout(guides = 'collect') 
prop_fig

residual_fig <-  pc_residual_biomass_by_mass+ pc_residual_abund_by_mass + plot_layout(guides = 'collect') 
residual_fig
```

```{r}
all_together <- prop_fig / residual_fig + plot_annotation(tag_levels = 'A') &
    theme(plot.tag = element_text(face = 'bold'))
all_together
```

```{r}
#ggsave("figures/Pcod_biomass_abund_w_residuals.png", plot = all_together, width = 8.5, height = 6, dpi = 300)
```
